Frtizing files for interested parties
The CA3083 was not at all good as it tends to require voltage boosting which could be done easier than shown here but there's more than one way to skin a cat as the saying goes and this one has many disadvantages anyway and as such was not used. Namely the need for up to 31 inputs. Consideration on multiplexing as a future addition. Overall system with code had sub-300ms delay for playing sound. However I introduced a 100ms delay in order to ensure that the history of notes playback on screen worked OK and potentially this could be reduced to sub-200ms without any induced delay.